+++ MyPatient.Models/Doctor_SurgicalProcedure.cs
@@ -8,7 +8,7 @@ namespace MyPatient.Models
 {
     public class Doctor_SurgicalProcedure
     {
-        public int DoctorId { get; set; }
+        public long DoctorId { get; set; }
         public Doctor? Doctor { get; set; }
 
         public long SurgicalProdecureId { get; set; }
